Supress the compiler error for elua_uip.c for now. This file has references to luaL_xxx functions.
Now, we have to figure out a way to get the code
to work for PicoC (and all the other languages
Alcor will support). This might be a bit of work.
We'll get back. For now, the code for PicoC on Alcor
compiles without issues (without uip).

Let's check if this actually works on the Mizar32.