Two changes to SConstruct
1. MatchEnumVariable('boot', ... ) now depends on the selected
language. Lua has two options - standard and luarpc. Until we
implement an RPC for PicoC, 'standard' will be the only boot
option for PicoC.

2. Ensure complete isolation between Lua and PicoC. Although
they share the code base, Lua specific headers/files should not
be seen by PicoC and inverse. This way, it is also much easier
to mark the features not common for both (for example elua_int).

Use #ifdef's in platform modules to decide what files/headers/
data structures to include for a selected language.