rename avr32's bool type to avr32_bool.
This avoids conflict with the bool defined
by PicoLisp.

Note: This may or may not be the best way
to go. When I just quickly looked up the
places where avr32's bool was used, I only
saw functions in the i2c block using this
type. However, PicoLisp (which uses it as
an enum type) has used this in several
places.

We have to get back to this if we have
other bools defined in other Alcor6L targets.