get lcdh_prin() to work with plisp_lcd_definechar().
The original implementation of lcdh_prin() will treat
numbers as strings and print them out on the LCD.
For user defined characters (0-7), this is not what is
needed.

The outNum_lcd() helper function has been commented
for now and a send_data function call has been used
in its place.