-falign-functions itself doesn't solve the problem of "unaligned functions".
surprising this trick with modifications to
pico.h and main.c is not required for avr32-gcc.
It just works on Mizar32 with -falign-functions.

We'll have to get back to this one later. For now,
we leave the changes in pico.h and main.c for all
platforms except Mizar32 (avr32). This needs more
investigation.